% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.fumeric}
\alias{as.fumeric}
\title{as factor then numeric}
\usage{
as.fumeric(x, levels = unique(x))
}
\arguments{
\item{x}{any vector}

\item{levels}{optional}
}
\description{
converts to factors then to numeric
}
\examples{
as.fumeric(c("a","x","x","y"))
}

