% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesort.R
\name{imagesort}
\alias{imagesort}
\title{image with sorted rows}
\usage{
imagesort(x, col = c("white", "black"), ...)
}
\arguments{
\item{x}{a matrix of 0s and 1s}

\item{col}{the colors of 0 and 1}

\item{...}{arguments to heatmap}
}
\description{
the rows are sorted such that the first column has 2 blocks,
the second column has 4 blocks, etc. see example("imagesort")
}
\examples{

x <- replicate(4,sample(0:1,40,TRUE))
imagesort(x)

}
\author{
Michael I. Love
}
