% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bioc.R
\name{install_bioc}
\alias{install_bioc}
\title{Install or update Bioconductor and CRAN packages}
\usage{
install_bioc(...)
}
\arguments{
\item{...}{arguments passed on to \code{link[BiocManager]{install}}}
}
\description{
This function is simply a wrapper for \code{link[BiocManager]{install}}. If \code{BiocManager} is not installed it it automatically installed.
}
\details{
If \code{BiocManager} is installed you can simply call \code{BiocManager::install} instead.
}
\examples{
\donttest{
  install_bioc("affy")
}
}
\author{
Rafael A. Irizarry
}
