% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/largeobj.R
\name{largeobj}
\alias{largeobj}
\title{What are the largest objects in memory?}
\usage{
largeobj(n = 5, units = "Mb")
}
\arguments{
\item{n}{the number of objects to return}

\item{units}{units to display, see \code{?object.size}}
}
\value{
a named character string of the size of the 'n' largest objects
}
\description{
This function lists all the objects in the global environmnet and lists the \code{n} largest.
}
\author{
Michael I. Love
}

