% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mypar.R
\name{mypar}
\alias{mypar}
\title{mypar}
\usage{
mypar(a = 1, b = 1, brewer.n = 8, brewer.name = "Dark2", cex.lab = 2,
  cex.main = 2, cex.axis = 1.5, mar = c(5.1, 5.1, 3.5, 2.1), mgp = c(3,
  1, 0), ...)
}
\arguments{
\item{a}{number of rows of plots}

\item{b}{number of columns of plots}

\item{...}{as in par()}
}
\description{
runs \code{\link{par}} with better settings (larger text)
}
\examples{
mypar(1,1)
plot(1,1,xlab="caca",ylab="Caca",main="Caca")
}

