% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullplot.R
\name{nullplot}
\alias{nullplot}
\title{nullplot}
\usage{
nullplot(x1 = 0, x2 = 1, y1 = 0, y2 = 1, xlab = "", ylab = "", ...)
}
\arguments{
\item{x1}{lowest x-axis value}

\item{x2}{largest x-axis value}

\item{y1}{lowest y-axis value}

\item{y2}{largest y-axis value}

\item{xlab}{x-axis title, defaults to no title}

\item{ylab}{y-axis title, defaults to no title}

\item{...}{further arguments passed on to plot}
}
\description{
Make an plot with nothing in it
}
\examples{
nullplot()
}
