% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{peek at the top of a text file}
\usage{
peek(x, n = 2)
}
\arguments{
\item{x}{a filename}

\item{n}{the number of lines to return}
}
\description{
this returns a character vector which shows the top n lines of a file
}
\author{
Michael I. Love
}

