% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sboxplot}
\alias{sboxplot}
\title{smart boxplot}
\usage{
sboxplot(x, ...)
}
\arguments{
\item{x}{a named list of numeric vectors}
}
\description{
draws points or boxes depending on sample size
}
\examples{
sboxplot(list(a=rnorm(15),b=rnorm(75),c=rnorm(10000)))
}

