% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shist.R
\name{shist}
\alias{shist}
\title{smart histogram}
\usage{
shist(z, unit, bw = "nrd0", n, from, to, plotHist = FALSE, add = FALSE,
  xlab, ylab = "Frequency", xlim, ylim, main, ...)
}
\arguments{
\item{z}{the data}

\item{unit}{the unit which determines the y axis scaling and is drawn}

\item{bw}{arguments to density}

\item{n}{arguments to density}

\item{from}{arguments to density}

\item{to}{arguments to density}

\item{...}{arguments to lines}
}
\description{
a smooth histogram with unit indicator (we're simply scaling the kernel density estimate)
}
\examples{
set.seed(1)
x = rnorm(50)
par(mfrow=c(2,1))
hist(x, breaks=-5:5)
shist(x, unit=1, xlim=c(-5,5))
}

