% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitit.R
\name{splitit}
\alias{splitit}
\title{split it}
\usage{
splitit(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Creates an list of indexes for each unique entry of \code{x}
}
\examples{
x <- c("a","a","b","a","b","c","b","b")
splitit(x)
}
