% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripplot.R
\name{stripplot}
\alias{stripplot}
\title{Better defaults for stripchart}
\usage{
stripplot(...)
}
\arguments{
\item{...}{passed to \code{stripchart}}
}
\value{
a plot
}
\description{
This simply calls \code{stripchart} but specifies
a vertical plot with jitter and using \code{pch=1}.
}
