% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whatMethods.R
\name{whatMethods}
\alias{whatMethods}
\title{What generics are defined for this S4 object?}
\usage{
whatMethods(object)
}
\arguments{
\item{object}{either an object, or a character specifying
the name of a class}
}
\value{
a character vector of generics defined for the
class of the object
}
\description{
This is a very beta prototype, and it only works with S4 objects.
}
\examples{
# from Advanced R: OO essentials material
# http://adv-r.had.co.nz/OO-essentials.html

library(stats4)
y <- c(26, 17, 13, 12, 20, 5, 9, 8, 5, 4, 8)
nLL <- function(lambda) - sum(dpois(y, lambda, log = TRUE))
fit <- mle(nLL, start = list(lambda = 5), nobs = length(y))
whatMethods(fit)
}
\references{
the bulk of this code was found with help from Gabe Becker within
\code{promptClass}
}

