% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fumeric.R
\name{as.fumeric}
\alias{as.fumeric}
\title{converts to factor and then numeric}
\usage{
as.fumeric(x, levels = unique(x))
}
\arguments{
\item{x}{a character vector}

\item{levels}{the leves to be used in the call to factor}
}
\description{
Converts a vector of characters into factors and then converts these into numeric.
}
\examples{

group = c("a","a","b","b")
plot(seq_along(group),col=as.fumeric(group))

}
\author{
Rafael A. Irizarry
}
