% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bartab.R
\name{bartab}
\alias{bartab}
\title{bartab}
\usage{
bartab(x, y, z, names, skipNone = FALSE, ...)
}
\arguments{
\item{x}{logical}

\item{y}{logical}

\item{z}{logical}

\item{names}{a character vector of length 3}

\item{skipNone}{remove the "none" group}

\item{...}{further arguments passed on to \code{\link{barplot}}}
}
\description{
Plot the overlap of three groups with a barplot
}
\author{
Michael I. Love
}

