% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imagemat.R
\name{imagemat}
\alias{imagemat}
\title{image of a matrix}
\usage{
imagemat(x, col = colorRampPalette(c("white", "black"))(9), las = 1,
  xlab = "", ylab = "", ...)
}
\arguments{
\item{x}{the matrix}

\item{col}{the colors}

\item{las}{as in par}

\item{...}{arguments passed to image}
}
\description{
produces an image of a matrix which matches
the natural orientation
}
\examples{
x <- matrix(c(1,0,0,0,1,
              1,1,0,1,1,
              1,0,1,0,1,
              1,0,0,0,1,
              1,0,0,0,1),
ncol=5,byrow=TRUE)

imagemat(x)
}

