% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bioc.R
\name{install_bioc}
\alias{install_bioc}
\title{Install or update Bioconductor and CRAN packages}
\usage{
install_bioc(...)
}
\arguments{
\item{...}{arguments passed on to \code{biocLite}}
}
\description{
This is function simply a wrapper for \code{biocLite}. It first sources the code 
from the Bioconductor site then calls \code{biocLite}.
}
\details{
Note that once you run this function in a session, you no
longer need to call since 
you can call \code{biocLite} directly.
}
\examples{
install_bioc("affy")
}
\author{
Rafael A. Irizarry
}
