% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{peek at the top of a text file}
\usage{
peek(x, n = 2)
}
\arguments{
\item{x}{a filename}

\item{n}{the number of lines to return}
}
\description{
this returns a character vector which shows the top n lines of a file.
Note: I realized after the fact that this is essentially a duplicate
of the base R function \code{readLines}.
}
\examples{

filename <- tempfile()
x<-matrix(round(rnorm(10^4),2),1000,10)
colnames(x)=letters[1:10]
write.csv(x,file=filename,row.names=FALSE)
peek(filename)
}
\author{
Michael I. Love
}
