% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popvar.R
\name{popvar}
\alias{popvar}
\title{population variance}
\usage{
popvar(x, ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{...}{further arguments passed along to \code{\link{var}}}
}
\description{
Returns the population variance. Note that \code{\link{var}} returns
the unbiased sample estimate of the population varaince. 
It simply multiplies the result of \code{\link{var}} by (n-1)  / n with n 
the populaton size.
}
\examples{
x <- c(0,1) ##variance should be 0.5^2=0.25
var(x) 
popvar(x)
}
