% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qcut.R
\name{qcut}
\alias{qcut}
\title{quantile cut}
\usage{
qcut(x, n, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{the number of quantiles}

\item{...}{further arguments passed on to \code{\link{cut}}}
}
\value{
a factor with n levels, from cutting x into quantiles
(0/n, 1/n, 2/n, ..., 1) of x. the first and last quantile have
1 subtracted and added respectively so all data falls into the bins.
}
\description{
\code{qcut} calls \code{\link{cut}} with the quantiles of \code{x} as breaks.
}
\examples{
qcut(rnorm(10), 4)
}
\seealso{
\code{\link[lsr]{quantileCut}} from the lsr package and
\code{\link[classInt]{classIntervals}} from the classInt package
}

