% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sboxplot.R
\name{sboxplot}
\alias{sboxplot}
\title{smart boxplot}
\usage{
sboxplot(x, ...)
}
\arguments{
\item{x}{a named list of numeric vectors}

\item{...}{further arguments passed on to \code{\link{boxplot}}}
}
\description{
draws points or boxes depending on sample size
}
\examples{

sboxplot(list(a=rnorm(15),b=rnorm(75),c=rnorm(10000)))

}
